requirements.txt
Pyrofork==2.3.68
fastapi==0.119.0
pymongo==4.15.3
python-dotenv==1.1.1
requests==2.32.5
TgCrypto==1.2.5
uvicorn==0.37.0
diskcache

